Simplify ZRoundResultConverter.
